feat: add a method to initialize normalizedBuckets for a given client. It sets the normalizedBuckets to buckets.
this method also removes the client's key from normalizedBucekts so that the PIR output does not include the client's own key. We don't want to remove the key altogether from the RT (buckets).

To avoid code duplication, we can make a removeKey (private) method that implements the logic in RemoveKey, without locking. RemoveKey and a RemoveKeyFromNormalizedBuckets can then lock the appropriate structure. This will require manually resolving a merge with main later on.